requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
biopython==1.80
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
coverage==7.0.5
cryptography==39.0.0
cycler==0.11.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.88.0
fonttools==4.38.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.2
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
mock==1.0.1
numpy==1.21.6
orjson==3.8.3
packaging==22.0
passlib==1.7.4
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pyparsing==3.0.9
pysam==0.20.0
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
RUST==0.1.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.11.0
